requirements.txt
Django==2.0.3
django-dotenv==1.4.2
django-jet==1.0.7
django-split-settings==0.3.0
et-xmlfile==1.0.1
jdcal==1.4
mysqlclient==1.3.12
openpyxl==2.5.4
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py-dateutil==2.2
pytz==2018.3
six==1.11.0
